+++ examples/demo-simple-stf/tests/stf_test.rs
@@ -2,6 +2,7 @@ use demo_simple_stf::{ApplySlotResult, CheckHashPreimageStf};
 use sov_mock_da::verifier::MockDaSpec;
 use sov_mock_da::{MockAddress, MockBlob, MockBlockHeader, MockValidityCond};
 use sov_mock_zkvm::MockZkVerifier;
+use sov_rollup_interface::services::da::RelevantBlobIters;
 use sov_rollup_interface::stf::StateTransitionFunction;
 
 #[test]
@@ -11,7 +12,7 @@ fn test_stf_success() {
     let stf = &mut CheckHashPreimageStf::<MockValidityCond>::default();
     StateTransitionFunction::<MockZkVerifier, MockDaSpec>::init_chain(stf, (), ());
 
-    let mut blobs = {
+    let mut batch_blobs = {
         let incorrect_preimage = vec![1; 32];
         let correct_preimage = vec![0; 32];
 
@@ -22,9 +23,25 @@ fn test_stf_success() {
     };
 
     // Pretend we are in native code and progress the blobs to the verified state.
-    for blob in &mut blobs {
+    for blob in &mut batch_blobs {
         blob.advance();
     }
+
+    let mut proof_blobs = {
+        [
+            MockBlob::new(vec![0; 32], address, [0; 32]),
+            MockBlob::new(vec![0; 32], address, [0; 32]),
+        ]
+    };
+
+    for blob in &mut proof_blobs {
+        blob.advance();
+    }
+
+    let relevant_blobs = RelevantBlobIters {
+        proof_blobs: &mut proof_blobs,
+        batch_blobs: &mut batch_blobs,
+    };
 
     let result = StateTransitionFunction::<MockZkVerifier, MockDaSpec>::apply_slot(
         stf,
@@ -33,7 +50,7 @@ fn test_stf_success() {
         (),
         &MockBlockHeader::default(),
         &MockValidityCond::default(),
-        &mut blobs,
+        relevant_blobs,
     );
 
     assert_eq!(2, result.batch_receipts.len());